feat(query): load ndjson support ts with diff units. (#17203)
* feat(query): load ndjson/csv/tsv support ts with diff units.

* fix

* refactor

* refactor

* fix clippy
