+++ tests/data/csv/ts.csv
@@ -1,0 +1,8 @@
+1736305864,('1736305864')
+1736305865000,('1736305865000')
+1736305866000000,('1736305866000000')
+1,('1')
+1970-01-01 00:00:01,('1970-01-01 00:00:01')
+1970-01-01 00:00:00.001,('1970-01-01 00:00:00.001')
+1970-01-01 00:00:00.000001,('1970-01-01 00:00:00.000001')
+1970-01-01 00:00:00.000000,('1970-01-01 00:00:00.000000')
