+++ tests/data/ndjson/ts.ndjson
@@ -1,0 +1,3 @@
+{"t":1736305864}
+{"t":1736305865000}
+{"t":1736305866000000}
