+++ tests/sqllogictests/suites/stage/formats/csv/csv_types.test
@@ -121,3 +121,23 @@ query
 select $1 from @csv_types(file_format => 'csv_raw');
 ----
  '((''c      d'',''{"a":1}''),{1:''["a",0]''},[({''x'':2},3)])'|1
+
+statement ok
+CREATE OR REPLACE TABLE cast_ts_csv (t timestamp, tt tuple(a timestamp))
+
+query
+copy into cast_ts_csv from @data/csv/ts.csv file_format = (type = CSV) ON_ERROR=continue
+----
+csv/ts.csv 8 0 NULL NULL
+
+query
+select t, tt.1 from cast_ts_csv
+----
+2025-01-08 03:11:04.000000 2025-01-08 03:11:04.000000
+2025-01-08 03:11:05.000000 2025-01-08 03:11:05.000000
+2025-01-08 03:11:06.000000 2025-01-08 03:11:06.000000
+1970-01-01 00:00:01.000000 1970-01-01 00:00:01.000000
+1970-01-01 00:00:01.000000 1970-01-01 00:00:01.000000
+1970-01-01 00:00:00.001000 1970-01-01 00:00:00.001000
+1970-01-01 00:00:00.000001 1970-01-01 00:00:00.000001
+1970-01-01 00:00:00.000000 1970-01-01 00:00:00.000000
