+++ tests/sqllogictests/suites/stage/formats/ndjson/ndjson_cast.test
@@ -15,3 +15,18 @@ select * from cast_ndjson
 data1 {'env':'test1','length':'ok'}
 data2 {'env':'test2','length':'true'}
 data3 {'env':'test3','length':'10'}
+
+statement ok
+CREATE OR REPLACE TABLE cast_ts_ndjson (t timestamp)
+
+query
+copy into cast_ts_ndjson from @data/ndjson/ts.ndjson file_format = (type = NDJSON) ON_ERROR=continue
+----
+ndjson/ts.ndjson 3 0 NULL NULL
+
+query
+select * from cast_ts_ndjson order by t
+----
+2025-01-08 03:11:04.000000
+2025-01-08 03:11:05.000000
+2025-01-08 03:11:06.000000
